Move datatype tests out of the main crate
